Fix infinite reload.
When the scrollbar appears and the content rectangle shrinks enough to decrease the number of columns, the images are reloaded. During the reload, the scrollbar disappears and the content rectangle grows enough to increase the number of columns and so on...
The space for the vertical scrollbar is now always reserved.